requirements.txt
python-telegram-bot==21.6
openai>=1.40.0
python-dotenv>=1.0.1
httpx>=0.27.2
